Apply linting to browser-provider code
For some reason the pre-commit hooks were broken on my machine.
This applies the fixes that were not applied before.

+++ packages/browser-provider/manual-test/metacoin-truffle/contracts/ConvertLib.sol
@@ -6,7 +6,10 @@ library ConvertLib {
         uint256 name;
     }
 
-    function convert(uint amount,uint conversionRate) public pure returns (uint convertedAmount)
+    function convert(uint256 amount, uint256 conversionRate)
+        public
+        pure
+        returns (uint256 convertedAmount)
     {
         return amount * conversionRate;
     }
